requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.0.0
chardet==3.0.4
discord.py==1.3.1
dnspython==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
httplib2==0.17.0
idna==2.8
multidict==4.7.4
mysql-connector-python==8.0.19
protobuf==3.6.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.11.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
websockets==8.1
yarl==1.4.2
